+++ lesson-6/query-side/src/main/java/com/edi/learn/axon/entries/ProductEntry.java
@@ -43,7 +43,7 @@ public class ProductEntry {
         this.name = name;
     }
 
-    public long getStock() {
+    public int getStock() {
         return stock;
     }
